Locomotion Module
Added a function which can be used with `pk_pills.custom.LocoAnim(ply,ent,"anim"`
It allows for a morphed player to use a model's animations and move along with it using locomotion data.

Additionally I added a small modification to `helpers.melee` which doesn't force an animation to play and has a customizable damage type, primarily for use with attacks that strike more than once.